+++ lib/provider/drift/my_user.dart
@@ -225,7 +225,6 @@ extension _MyUserDb on DtoMyUser {
             e.muted == null
                 ? null
                 : MuteDuration.fromJson(jsonDecode(e.muted!)),
-        blocklistCount: e.blocklistCount,
         lastSeenAt: e.lastSeenAt,
         welcomeMessage:
             e.welcomeMessage == null
@@ -261,7 +260,6 @@ extension _MyUserDb on DtoMyUser {
       presenceIndex: value.presenceIndex,
       online: value.online,
       muted: value.muted == null ? null : jsonEncode(value.muted?.toJson()),
-      blocklistCount: value.blocklistCount,
       lastSeenAt: value.lastSeenAt,
       ver: ver.val,
       welcomeMessage: