+++ test/widget/chat_edit_message_test.dart
@@ -252,11 +252,12 @@ void main() async {
   when(
     graphQlProvider.favoriteChatsEvents(any),
   ).thenAnswer((_) => const Stream.empty());
-
+  when(
+    graphQlProvider.blocklistEvents(any),
+  ).thenAnswer((_) => const Stream.empty());
   when(
     graphQlProvider.myUserEvents(any),
   ).thenAnswer((_) async => const Stream.empty());
-
   when(
     graphQlProvider.sessionsEvents(any),
   ).thenAnswer((_) => const Stream.empty());
@@ -387,7 +388,6 @@ void main() async {
         blocklistProvider,
         userRepository,
         versionProvider,
-        myUserProvider,
         me: const UserId('me'),
       ),
     );