+++ test/widget/chat_reply_message_test.dart
@@ -311,15 +311,15 @@ void main() async {
   when(
     graphQlProvider.incomingCallsTopEvents(3),
   ).thenAnswer((_) => const Stream.empty());
-
   when(
     graphQlProvider.favoriteChatsEvents(any),
   ).thenAnswer((_) => const Stream.empty());
-
+  when(
+    graphQlProvider.blocklistEvents(any),
+  ).thenAnswer((_) => const Stream.empty());
   when(
     graphQlProvider.myUserEvents(any),
   ).thenAnswer((_) async => const Stream.empty());
-
   when(
     graphQlProvider.sessionsEvents(any),
   ).thenAnswer((_) => const Stream.empty());
@@ -437,7 +437,6 @@ void main() async {
         blocklistProvider,
         userRepository,
         versionProvider,
-        myUserProvider,
         me: const UserId('me'),
       ),
     );